Only allow user to delete own notes

+++ app/controllers/notes_controller.rb
@@ -4,23 +4,26 @@ class NotesController < ApplicationController
   def create
     note = Note.new(note_params)
     note.user_id = current_user.id
-    note.track_id = params[:id]
+    note.track_id = params[:track_id]
 
-    flash[:error] = 'Unable to save note, please try again' unless note.save
+    flash[:error] = 'Note cannot be blank!' unless note.save
     redirect_to request.referrer
   end
 
   def destroy
     note = Note.find_by_id(params[:id])
-    if note
+
+    if note && current_user.id == note.user_id
       note.destroy
+      redirect_to request.referrer
+    else
+      render plain: 'Forbidden, you cannot do that!', status: :forbidden
     end
-    redirect_to request.referrer
   end
 
   private
 
   def note_params
-    params.require(:note).permit(:user_id, :track_id, :body)
+    params.require(:note).permit(:track_id, :body)
   end
 end
